object 220713 17:07

+++ src/com/iu/object1/ex1/Student.java
@@ -7,6 +7,6 @@ public class Student {
 	int eng;
 	int math;
 	int total;
-	int avg;
+	double avg;
 
 }
